Move comparison function objects Less and Greater out of <ArborX_DetailsPriorityQueue.hpp>
